make serial driver shim for EPD
I hadn't noticed when implementing CONSOLE first, but the whole
escape/input system handling is implemented in `uart_console.c`, so apart
from re-implementing it, the easiest way is writing a UART driver.

That's what this commit does.
The input is still stubbed, we could think about grabbing chars from the
USB UART as a first step.